move `_get_fn_args_from_batch` to utils
