+++ pyrovelocity/_velocity_model_minimal.py
@@ -348,49 +348,3 @@ class VelocityModelAuto(LogNormalModel):
                 u = pyro.sample("u", u_dist, obs=u_obs)
                 s = pyro.sample("s", s_dist, obs=s_obs)
         return u, s
-
-
-# TODO: move to utils or VelocityModule
-# @staticmethod
-# def _get_fn_args_from_batch(
-#     tensor_dict: Dict[str, torch.Tensor]
-# ) -> Tuple[
-#     Tuple[
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         torch.Tensor,
-#         None,
-#         None,
-#     ],
-#     Dict[Any, Any],
-# ]:
-#     u_obs = tensor_dict["U"]
-#     s_obs = tensor_dict["X"]
-#     u_log_library = tensor_dict["u_lib_size"]
-#     s_log_library = tensor_dict["s_lib_size"]
-#     u_log_library_mean = tensor_dict["u_lib_size_mean"]
-#     s_log_library_mean = tensor_dict["s_lib_size_mean"]
-#     u_log_library_scale = tensor_dict["u_lib_size_scale"]
-#     s_log_library_scale = tensor_dict["s_lib_size_scale"]
-#     ind_x = tensor_dict["ind_x"].long().squeeze()
-#     cell_state = tensor_dict.get("pyro_cell_state")
-#     time_info = tensor_dict.get("time_info")
-#     return (
-#         u_obs,
-#         s_obs,
-#         u_log_library,
-#         s_log_library,
-#         u_log_library_mean,
-#         s_log_library_mean,
-#         u_log_library_scale,
-#         s_log_library_scale,
-#         ind_x,
-#         cell_state,
-#         time_info,
-#     ), {}
